requirements.txt
PyQt>=5.9.2
pandas>=1.0.5
networkx>=2.4
matplotlib>=3.2.2
setuptools>=46.4.0
pyodbc>=4.0.30